♻️ ✅ Refactor unit test architecture
Previously, we were running our unit tests with the architecture:

jest -> docker -> nyc -> ts-node -> action

This had a number of problems or quirks, including having to combine all
the individual istanbul (nyc) coverage results into one report,
and generally being harder to debug.

The architecture has now been changed to:

bash script -> docker -> jest

greatly simplifying the structure, removing the need for coverage
processing. To facilitate this, the main action code had to be reworked
to allow it to be run by a function call (and not just when a module
is imported).